app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
boto3==1.34.69
